refactored tests for HeuristicPermutationGenerator
